+++ src/network/message.rs
@@ -4,7 +4,7 @@ use chrono::prelude::*;
 use serde_derive::{Deserialize, Serialize};
 use strum_macros::{Display, EnumString};
 
-use crate::state::{GroupStatus, State};
+use crate::state::{Group, State};
 use crate::task::Task;
 
 /// This is the main message enum. \
@@ -178,8 +178,7 @@ pub enum GroupMessage {
 
 #[derive(PartialEq, Clone, Debug, Deserialize, Serialize)]
 pub struct GroupResponseMessage {
-    pub groups: BTreeMap<String, GroupStatus>,
-    pub settings: BTreeMap<String, usize>,
+    pub groups: BTreeMap<String, Group>,
 }
 
 #[derive(PartialEq, Clone, Debug, Deserialize, Serialize)]