+++ src/settings.rs
@@ -1,4 +1,3 @@
-use std::collections::{BTreeMap, HashMap};
 use std::fs::{create_dir_all, File};
 use std::io::prelude::*;
 use std::path::{Path, PathBuf};
@@ -10,8 +9,6 @@ use shellexpand::tilde;
 
 use crate::error::Error;
 use crate::platform::directories::*;
-
-pub const PUEUE_DEFAULT_GROUP: &str = "default";
 
 /// All settings which are used by both, the client and the daemon
 #[derive(PartialEq, Clone, Debug, Deserialize, Serialize)]
@@ -91,9 +88,6 @@ pub struct Daemon {
     pub callback: Option<String>,
     /// The amount of log lines from stdout/stderr that are passed to the callback command.
     pub callback_log_lines: usize,
-    /// This shouldn't be manipulated manually if the daemon is running.
-    /// This represents all known groups and their amount of parallel tasks.
-    pub groups: BTreeMap<String, usize>,
 }
 
 /// The parent settings struct. \
@@ -135,12 +129,7 @@ impl Settings {
     /// If no config files can be found or fields are missing, an error is returned.
     pub fn read(from_file: &Option<PathBuf>) -> Result<Settings, Error> {
         let config = Config::new();
-
-        // Insert the default group, if it doesn't exist.
-        let mut settings = parse_config(config, true, from_file)?;
-        settings.ensure_default_group();
-
-        Ok(settings)
+        parse_config(config, true, from_file)
     }
 
     /// Try to read existing config files and
@@ -159,11 +148,7 @@ impl Settings {
         from_file: &Option<PathBuf>,
     ) -> Result<Settings, Error> {
         let config = Settings::default_config()?;
-
-        let mut settings = parse_config(config, require_config, from_file)?;
-        settings.ensure_default_group();
-
-        Ok(settings)
+        parse_config(config, require_config, from_file)
     }
 
     pub fn default_config() -> Result<Config, Error> {
@@ -240,9 +225,6 @@ impl Settings {
             .set_default("daemon.callback", None::<String>)
             .unwrap();
         config.set_default("daemon.callback_log_lines", 10).unwrap();
-        config
-            .set_default("daemon.groups", HashMap::<String, i64>::new())
-            .unwrap();
 
         Ok(config)
     }
@@ -278,15 +260,6 @@ impl Settings {
         file.write_all(content.as_bytes())?;
 
         Ok(())
-    }
-
-    pub fn ensure_default_group(&mut self) {
-        // Insert the default group, if it doesn't exist.
-        if self.daemon.groups.get(PUEUE_DEFAULT_GROUP).is_none() {
-            self.daemon
-                .groups
-                .insert(PUEUE_DEFAULT_GROUP.to_string(), 1);
-        }
     }
 }